Added options to quickly disable script and handle images wrongly affected by aspect ratio change.
These are really lazy changes.

If you would like to disable the script, simply change ratioFix value to false.

I also included a fix for images that could be included in some video playlists/folders. It works so that if it detects an image, it would reset it to its original aspect ratio, not the previous video file aspect ratio change.

To disable it, simply change imagesFix value to false.

Reason for images fix:
The logic in this aspect ratio change is really weird. For example, consider the following playlist:

- video_001.mp4 (16:9) [unaffected by script]
- youtube.com/video [unaffected by script]
- video_002.mp4 (4:3) [affected by script, ratio changed to 16:9]
- image_002.jpg [affected by mpv because it retains the previous aspect ratio]
- image_003.jpg [same as previous image]
- video_003.mp4 (not 4:3) [aspect ratio reset, everything back to normal]
... So on, so forth.

So if an image is in the middle of the playlist and it happens to be right after a video that was affected by the script, all the images will have that 16:9 aspect ratio UNTIL another video plays.

So, this lazy method fixes that issue.

This also made me realize that the script could be much more efficient, in fact, in something other than LUA I could probably short it all down to a single line, give or take. This is probably due to my lack of knowledge though, not LUA's fault.

I either need to rethink the logic, or go through MPV's parameters and options once over to see if there are better aspect ratio handling, or both! One thing I'm hoping for MPV to do is to include a simple MIME-type check, instead of this file extension handling.